requirements.txt
fastapi[standard]==0.115.6

# Geospatial processing libraries
Pillow==10.0.1
pyproj==3.6.0
lxml==5.3.0
rasterio==1.4.3
pydantic>=2.7.0,<3.0.0
pydantic-settings==2.3.4
sqlalchemy==2.0.31
alembic==1.13.2
rio-cogeo==5.4.1